Move files related to system assembly into the system folder
