+++ src/restruct_poc/solver/Step.hpp
@@ -12,7 +12,7 @@
 #include "UpdateAlgorithmicAcceleration.hpp"
 #include "UpdateStatePrediction.hpp"
 
-#include "src/restruct_poc/UpdateState.hpp"
+#include "src/restruct_poc/system/UpdateState.hpp"
 #include "src/restruct_poc/beams/Beams.hpp"
 
 namespace openturbine {
