+++ src/restruct_poc/system/AssembleElasticStiffnessMatrix.hpp
@@ -3,14 +3,14 @@
 #include <Kokkos_Core.hpp>
 #include <Kokkos_Profiling_ScopedRegion.hpp>
 
-#include "IntegrateMatrix.hpp"
+#include "IntegrateElasticStiffnessMatrix.hpp"
 #include "src/restruct_poc/beams/Beams.hpp"
-#include "types.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
 
-inline void AssembleInertialStiffnessMatrix(Beams& beams, View_NxN K) {
-    auto region = Kokkos::Profiling::ScopedRegion("Assemble Inertial Stiffness Matrix");
+inline void AssembleElasticStiffnessMatrix(Beams& beams, View_NxN K) {
+    auto region = Kokkos::Profiling::ScopedRegion("Assemble Elastic Stiffness Matrix");
     auto range_policy = std::invoke([&]() {
         if constexpr (std::is_same_v<
                           Kokkos::DefaultExecutionSpace, Kokkos::DefaultHostExecutionSpace>) {
@@ -23,14 +23,18 @@ inline void AssembleInertialStiffnessMatrix(Beams& beams, View_NxN K) {
         }
     });
     Kokkos::parallel_for(
-        "IntegrateMatrix", range_policy,
-        IntegrateMatrix{
+        "IntegrateElasticStiffnessMatrix", range_policy,
+        IntegrateElasticStiffnessMatrix{
             beams.elem_indices,
             beams.node_state_indices,
             beams.qp_weight,
             beams.qp_jacobian,
             beams.shape_interp,
-            beams.qp_Kuu,
+            beams.shape_deriv,
+            beams.qp_Puu,
+            beams.qp_Cuu,
+            beams.qp_Ouu,
+            beams.qp_Quu,
             K,
         }
     );
