+++ src/restruct_poc/system/AssembleGyroscopicInertiaMatrix.hpp
