+++ src/restruct_poc/system/AssembleInertialStiffnessMatrix.hpp
@@ -3,14 +3,14 @@
 #include <Kokkos_Core.hpp>
 #include <Kokkos_Profiling_ScopedRegion.hpp>
 
-#include "IntegrateElasticStiffnessMatrix.hpp"
+#include "IntegrateMatrix.hpp"
 #include "src/restruct_poc/beams/Beams.hpp"
-#include "types.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
 
-inline void AssembleElasticStiffnessMatrix(Beams& beams, View_NxN K) {
-    auto region = Kokkos::Profiling::ScopedRegion("Assemble Elastic Stiffness Matrix");
+inline void AssembleInertialStiffnessMatrix(Beams& beams, View_NxN K) {
+    auto region = Kokkos::Profiling::ScopedRegion("Assemble Inertial Stiffness Matrix");
     auto range_policy = std::invoke([&]() {
         if constexpr (std::is_same_v<
                           Kokkos::DefaultExecutionSpace, Kokkos::DefaultHostExecutionSpace>) {
@@ -23,18 +23,14 @@ inline void AssembleElasticStiffnessMatrix(Beams& beams, View_NxN K) {
         }
     });
     Kokkos::parallel_for(
-        "IntegrateElasticStiffnessMatrix", range_policy,
-        IntegrateElasticStiffnessMatrix{
+        "IntegrateMatrix", range_policy,
+        IntegrateMatrix{
             beams.elem_indices,
             beams.node_state_indices,
             beams.qp_weight,
             beams.qp_jacobian,
             beams.shape_interp,
-            beams.shape_deriv,
-            beams.qp_Puu,
-            beams.qp_Cuu,
-            beams.qp_Ouu,
-            beams.qp_Quu,
+            beams.qp_Kuu,
             K,
         }
     );
