+++ src/restruct_poc/system/CalculateInertiaStiffnessMatrix.hpp
@@ -5,8 +5,8 @@
 #include <KokkosBlas1_set.hpp>
 #include <Kokkos_Core.hpp>
 
-#include "VectorOperations.hpp"
-#include "types.hpp"
+#include "src/restruct_poc/VectorOperations.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
