+++ src/restruct_poc/system/CalculateMassMatrixComponents.hpp
@@ -2,8 +2,8 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "VectorOperations.hpp"
-#include "types.hpp"
+#include "src/restruct_poc/VectorOperations.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
