+++ src/restruct_poc/system/CalculateStrain.hpp
@@ -4,8 +4,8 @@
 #include <KokkosBlas.hpp>
 #include <Kokkos_Core.hpp>
 
-#include "QuaternionOperations.hpp"
-#include "types.hpp"
+#include "src/restruct_poc/QuaternionOperations.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
