+++ src/restruct_poc/system/CalculateTemporaryVariables.hpp
@@ -2,9 +2,8 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "VectorOperations.hpp"
-#include "types.hpp"
-#include "VectorOperations.hpp"
+#include "src/restruct_poc/VectorOperations.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
