+++ src/restruct_poc/system/IntegrateMatrix.hpp
@@ -2,23 +2,19 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "beams/Beams.hpp"
-#include "types.hpp"
+#include "src/restruct_poc/beams/Beams.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
 
-struct IntegrateElasticStiffnessMatrix {
+struct IntegrateMatrix {
     Kokkos::View<Beams::ElemIndices*>::const_type elem_indices;  // Element indices
     Kokkos::View<int*>::const_type node_state_indices;           // Element indices
     View_N::const_type qp_weight_;                               //
     View_N::const_type qp_jacobian_;                             // Jacobians
     View_NxN::const_type shape_interp_;                          // Num Nodes x Num Quadrature points
-    View_NxN::const_type shape_deriv_;                           // Num Nodes x Num Quadrature points
-    View_Nx6x6::const_type qp_Puu_;                              //
-    View_Nx6x6::const_type qp_Cuu_;                              //
-    View_Nx6x6::const_type qp_Ouu_;                              //
-    View_Nx6x6::const_type qp_Quu_;                              //
-    View_NxN_atomic gbl_M_;
+    View_Nx6x6::const_type qp_M_;                                //
+    View_NxN_atomic gbl_M_;                                      //
 
     KOKKOS_FUNCTION
     void operator()(const int i_elem) const {
@@ -30,31 +26,23 @@ struct IntegrateElasticStiffnessMatrix {
                 auto local_M = Kokkos::View<double[6][6], Kokkos::MemoryTraits<Kokkos::Unmanaged>>(
                     local_M_data.data()
                 );
-                for (int k = 0; k < idx.num_qps; ++k) {  // QPs
+                for (int k = 0; k < idx.num_qps; ++k) {
                     const auto k_qp = idx.qp_range.first + k;
                     const auto w = qp_weight_(k_qp);
                     const auto jacobian = qp_jacobian_(k_qp);
                     const auto phi_i = shape_interp_(i, k);
                     const auto phi_j = shape_interp_(j, k);
-                    const auto phi_prime_i = shape_deriv_(i, k);
-                    const auto phi_prime_j = shape_deriv_(j, k);
-                    const auto coeff_P = w * (phi_i * phi_prime_j);
-                    const auto coeff_Q = w * (phi_i * phi_j * jacobian);
-                    const auto coeff_C = w * (phi_prime_i * phi_prime_j / jacobian);
-                    const auto coeff_O = w * (phi_prime_i * phi_j);
-                    for (int m = 0; m < 6; ++m) {      // Matrix components
-                        for (int n = 0; n < 6; ++n) {  // Matrix components
-                            local_M(m, n) +=
-                                coeff_P * qp_Puu_(k_qp, m, n) + coeff_Q * qp_Quu_(k_qp, m, n) +
-                                coeff_C * qp_Cuu_(k_qp, m, n) + coeff_O * qp_Ouu_(k_qp, m, n);
+                    const auto coeff = w * phi_i * phi_j * jacobian;
+                    for (int m = 0; m < kLieAlgebraComponents; ++m) {
+                        for (int n = 0; n < kLieAlgebraComponents; ++n) {
+                            local_M(m, n) += coeff * qp_M_(k_qp, m, n);
                         }
                     }
                 }
-
                 const auto i_gbl_start = node_state_indices(i) * kLieAlgebraComponents;
                 const auto j_gbl_start = node_state_indices(j) * kLieAlgebraComponents;
-                for (int m = 0; m < 6; ++m) {
-                    for (int n = 0; n < 6; ++n) {
+                for (int m = 0; m < kLieAlgebraComponents; ++m) {
+                    for (int n = 0; n < kLieAlgebraComponents; ++n) {
                         gbl_M_(i_gbl_start + m, j_gbl_start + n) += local_M(m, n);
                     }
                 }
@@ -80,22 +68,14 @@ struct IntegrateElasticStiffnessMatrix {
                     const auto jacobian = qp_jacobian_(k_qp);
                     const auto phi_i = shape_interp_(i, k);
                     const auto phi_j = shape_interp_(j, k);
-                    const auto phi_prime_i = shape_deriv_(i, k);
-                    const auto phi_prime_j = shape_deriv_(j, k);
-                    const auto coeff_P = w * (phi_i * phi_prime_j);
-                    const auto coeff_Q = w * (phi_i * phi_j * jacobian);
-                    const auto coeff_C = w * (phi_prime_i * phi_prime_j / jacobian);
-                    const auto coeff_O = w * (phi_prime_i * phi_j);
+                    const auto coeff = w * phi_i * phi_j * jacobian;
                     Kokkos::parallel_for(
                         Kokkos::ThreadVectorMDRange(member, 6, 6),
                         [=](int m, int n) {
-                            local_M(m, n) +=
-                                coeff_P * qp_Puu_(k_qp, m, n) + coeff_Q * qp_Quu_(k_qp, m, n) +
-                                coeff_C * qp_Cuu_(k_qp, m, n) + coeff_O * qp_Ouu_(k_qp, m, n);
+                            local_M(m, n) += coeff * qp_M_(k_qp, m, n);
                         }
                     );
                 }
-
                 const auto i_gbl_start = node_state_indices(i) * kLieAlgebraComponents;
                 const auto j_gbl_start = node_state_indices(j) * kLieAlgebraComponents;
                 Kokkos::parallel_for(Kokkos::ThreadVectorMDRange(member, 6, 6), [=](int m, int n) {
@@ -124,24 +104,17 @@ struct IntegrateElasticStiffnessMatrix {
             const auto jacobian = qp_jacobian_(k_qp);
             const auto phi_i = shape_interp_(i, k);
             const auto phi_j = shape_interp_(j, k);
-            const auto phi_prime_i = shape_deriv_(i, k);
-            const auto phi_prime_j = shape_deriv_(j, k);
-            const auto coeff_P = w * (phi_i * phi_prime_j);
-            const auto coeff_Q = w * (phi_i * phi_j * jacobian);
-            const auto coeff_C = w * (phi_prime_i * phi_prime_j / jacobian);
-            const auto coeff_O = w * (phi_prime_i * phi_j);
-            for (int m = 0; m < 6; ++m) {      // Matrix components
-                for (int n = 0; n < 6; ++n) {  // Matrix components
-                    local_M(m, n) += coeff_P * qp_Puu_(k_qp, m, n) + coeff_Q * qp_Quu_(k_qp, m, n) +
-                                     coeff_C * qp_Cuu_(k_qp, m, n) + coeff_O * qp_Ouu_(k_qp, m, n);
+            const auto coeff = w * phi_i * phi_j * jacobian;
+            for (int m = 0; m < kLieAlgebraComponents; ++m) {
+                for (int n = 0; n < kLieAlgebraComponents; ++n) {
+                    local_M(m, n) += coeff * qp_M_(k_qp, m, n);
                 }
             }
         }
-
         const auto i_gbl_start = node_state_indices(i) * kLieAlgebraComponents;
         const auto j_gbl_start = node_state_indices(j) * kLieAlgebraComponents;
-        for (int m = 0; m < 6; ++m) {
-            for (int n = 0; n < 6; ++n) {
+        for (int m = 0; m < kLieAlgebraComponents; ++m) {
+            for (int n = 0; n < kLieAlgebraComponents; ++n) {
                 gbl_M_(i_gbl_start + m, j_gbl_start + n) += local_M(m, n);
             }
         }
@@ -149,7 +122,7 @@ struct IntegrateElasticStiffnessMatrix {
 
     KOKKOS_FUNCTION
     void operator()(const int i_elem, const int i_index, const int j_index, const int k) const {
-        const auto idx = elem_indices(i_elem);
+        const auto idx = elem_indices[i_elem];
 
         if (i_index >= idx.num_nodes || j_index >= idx.num_nodes || k >= idx.num_qps) {
             return;
@@ -165,23 +138,16 @@ struct IntegrateElasticStiffnessMatrix {
         const auto jacobian = qp_jacobian_(k_qp);
         const auto phi_i = shape_interp_(i, k);
         const auto phi_j = shape_interp_(j, k);
-        const auto phi_prime_i = shape_deriv_(i, k);
-        const auto phi_prime_j = shape_deriv_(j, k);
-        const auto coeff_P = w * (phi_i * phi_prime_j);
-        const auto coeff_Q = w * (phi_i * phi_j * jacobian);
-        const auto coeff_C = w * (phi_prime_i * phi_prime_j / jacobian);
-        const auto coeff_O = w * (phi_prime_i * phi_j);
-        for (int m = 0; m < 6; ++m) {      // Matrix components
-            for (int n = 0; n < 6; ++n) {  // Matrix components
-                local_M(m, n) = coeff_P * qp_Puu_(k_qp, m, n) + coeff_Q * qp_Quu_(k_qp, m, n) +
-                                coeff_C * qp_Cuu_(k_qp, m, n) + coeff_O * qp_Ouu_(k_qp, m, n);
+        const auto coeff = w * phi_i * phi_j * jacobian;
+        for (int m = 0; m < kLieAlgebraComponents; ++m) {
+            for (int n = 0; n < kLieAlgebraComponents; ++n) {
+                local_M(m, n) += coeff * qp_M_(k_qp, m, n);
             }
         }
-
         const auto i_gbl_start = node_state_indices(i) * kLieAlgebraComponents;
         const auto j_gbl_start = node_state_indices(j) * kLieAlgebraComponents;
-        for (int m = 0; m < 6; ++m) {
-            for (int n = 0; n < 6; ++n) {
+        for (int m = 0; m < kLieAlgebraComponents; ++m) {
+            for (int n = 0; n < kLieAlgebraComponents; ++n) {
                 gbl_M_(i_gbl_start + m, j_gbl_start + n) += local_M(m, n);
             }
         }
