+++ src/restruct_poc/system/IntegrateResidualVector.hpp
@@ -2,7 +2,7 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "types.hpp"
+#include "src/restruct_poc/types.hpp"
 
 namespace openturbine {
