+++ tests/unit_tests/restruct_poc/test_beams.cpp
@@ -2,17 +2,17 @@
 
 #include <gtest/gtest.h>
 
-#include "src/restruct_poc/AssembleElasticStiffnessMatrix.hpp"
-#include "src/restruct_poc/AssembleGyroscopicInertiaMatrix.hpp"
-#include "src/restruct_poc/AssembleInertialStiffnessMatrix.hpp"
-#include "src/restruct_poc/AssembleMassMatrix.hpp"
-#include "src/restruct_poc/AssembleResidualVector.hpp"
+#include "src/restruct_poc/system/AssembleElasticStiffnessMatrix.hpp"
+#include "src/restruct_poc/system/AssembleGyroscopicInertiaMatrix.hpp"
+#include "src/restruct_poc/system/AssembleInertialStiffnessMatrix.hpp"
+#include "src/restruct_poc/system/AssembleMassMatrix.hpp"
+#include "src/restruct_poc/system/AssembleResidualVector.hpp"
 #include "src/restruct_poc/beams/BeamElement.hpp"
 #include "src/restruct_poc/beams/BeamNode.hpp"
 #include "src/restruct_poc/beams/BeamSection.hpp"
 #include "src/restruct_poc/beams/BeamsInput.hpp"
 #include "src/restruct_poc/beams/CreateBeams.hpp"
-#include "src/restruct_poc/UpdateState.hpp"
+#include "src/restruct_poc/system/UpdateState.hpp"
 #include "src/restruct_poc/beams/Beams.hpp"
 #include "src/restruct_poc/solver/State.hpp"
 #include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
